[FIX] make new module inherit l10n_br_fiscal instead of l10n_br_nfe
