+++ l10n_br_fiscal_purchase/models/document.py
