+++ l10n_br_fiscal_purchase/readme/CONTRIBUTORS.rst
