+++ l10n_br_fiscal_purchase/views/document_view.xml
@@ -1,6 +1,8 @@
 <odoo>
-    <record id="nfe_document_form" model="ir.ui.view">
-        <field name="name">l10n_br_nfe.document.form (in l10n_br_nfe_purchase)</field>
+    <record id="fiscal_purchase_document_form" model="ir.ui.view">
+        <field
+            name="name"
+        >l10n_br_nfe.fiscal.purchase.document.form (in l10n_br_fiscal_purchase)</field>
         <field name="model">l10n_br_fiscal.document</field>
         <field name="inherit_id" ref="l10n_br_fiscal.document_form" />
         <field name="arch" type="xml">
@@ -16,7 +18,7 @@
                     <field
                         name="linked_purchase_count"
                         widget="statinfo"
-                        string="Ordens de Compra"
+                        string="Purchase Orders"
                     />
                 </button>
             </xpath>
