+++ l10n_br_nfe/__manifest__.py
@@ -13,7 +13,7 @@
     "development_status": "Beta",
     "version": "14.0.8.4.1",
     "depends": [
-        "l10n_br_fiscal",
+        "l10n_br_fiscal_purchase",
         "l10n_br_nfe_spec",
         "spec_driven_model",
     ],
