--- l10n_br_nfe_purchase/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizards
