--- l10n_br_nfe_purchase/wizards/__init__.py
@@ -1,1 +1,0 @@
-from . import import_document
