--- l10n_br_nfe_purchase/wizards/import_document.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2023 KMEE
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-<odoo>
-
-    <record id="l10n_br_nfe_import_xml_form" model="ir.ui.view">
-        <field name="name">l10n_br_nfe.import_xml.form (in l10n_br_nfe_purchase)</field>
-        <field name="model">l10n_br_nfe.import_xml</field>
-        <field name="inherit_id" ref="l10n_br_nfe.l10n_br_nfe_import_xml_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//group[@id='document_info']" position="after">
-                <separator
-                    string="Conciliar com uma ordem de compra"
-                    attrs="{'invisible': [('xml', '=', False)]}"
-                />
-                <group attrs="{'invisible': [('xml', '=', False)]}">
-                    <group>
-                        <field name="purchase_link_type" />
-                        <field
-                            name="purchase_id"
-                            attrs="{'invisible': [('purchase_link_type', '!=', 'choose')]}"
-                        />
-                    </group>
-                </group>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
