+++ setup/l10n_br_fiscal_purchase/odoo/addons/l10n_br_fiscal_purchase
@@ -1,0 +1,1 @@
+../../../../l10n_br_fiscal_purchase
