--- setup/l10n_br_nfe_purchase/odoo/addons/l10n_br_nfe_purchase
@@ -1,1 +1,0 @@
-../../../../l10n_br_nfe_purchase
